Refactored world generation into something more usable.

+++ Caligo.Core/Utils/Random.cs
@@ -40,8 +40,8 @@ using System;
 public class Random
 {
 // The +1 ensures NextDouble doesn't generate 1.0
-    const double REAL_UNIT_INT = 1.0 / ((double)int.MaxValue + 1.0);
-    const double REAL_UNIT_UINT = 1.0 / ((double)uint.MaxValue + 1.0);
+    const double REAL_UNIT_INT = 1.0 / (int.MaxValue + 1.0);
+    const double REAL_UNIT_UINT = 1.0 / (uint.MaxValue + 1.0);
     private const uint Y = 842502087, Z = 3579807591, W = 273326509;
 
     uint x, y, z, w;
@@ -162,14 +162,14 @@ public class Random
         {
             // If range is <0 then an overflow has occured and must resort to using long integer arithmetic instead (slower).
             // We also must use all 32 bits of precision, instead of the normal 31, which again is slower.	
-            return lowerBound + (int)(REAL_UNIT_UINT * (double)(w = w ^ w >> 19 ^ t ^ t >> 8) *
-                                      (double)((long)upperBound - (long)lowerBound));
+            return lowerBound + (int)(REAL_UNIT_UINT * (w = w ^ w >> 19 ^ t ^ t >> 8) *
+                                      (upperBound - (long)lowerBound));
         }
 
         // 31 bits of precision will suffice if range<=int.MaxValue. This allows us to cast to an int and gain
         // a little more performance.
         return lowerBound +
-               (int)(REAL_UNIT_INT * (double)(int)(0x7FFFFFFF & (w = w ^ w >> 19 ^ t ^ t >> 8)) * (double)range);
+               (int)(REAL_UNIT_INT * (int)(0x7FFFFFFF & (w = w ^ w >> 19 ^ t ^ t >> 8)) * range);
     }
 
     /// <summary>
